requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==4.2
django-credit-cards==0.4.1
django-countries==7.5.1
executing==1.2.0
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.3
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycryptodome==3.17
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
virtualenv==20.23.0
wcwidth==0.2.6
Werkzeug==2.3.3
